Improve role REST controller

+++ src/main/java/com/DTeam/eshop/controllers/RoleController.java
@@ -4,15 +4,21 @@ import java.util.List;
 
 import com.DTeam.eshop.entities.Role;
 import com.DTeam.eshop.services.RoleService;
+import com.DTeam.eshop.utilities.CustomErrorType;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
 
 @RestController
 @RequestMapping("/api")
@@ -20,29 +26,72 @@ public class RoleController {
 
     @Autowired private RoleService roleService;
 
+    //Retrieve all roles
     @GetMapping("/roles")
-    public List<Role> getRoles(){
-        return roleService.listAll();
+    public ResponseEntity<List<Role>> getRoles(){
+        List<Role> roles = roleService.listAll();
+        if(roles.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<List<Role>>(roles, HttpStatus.OK);
     }
 
-    @GetMapping("/role/{name}")
-    public Role getRole(@PathVariable String name){
-        return roleService.get(name);
+    //Retrieve single role
+	@GetMapping("/role/{name}")
+	public ResponseEntity<?> getRole(@PathVariable("name") String name) {
+        if(roleService.isRoleExist(name)){
+            Role role = roleService.get(name);
+            return new ResponseEntity<Role>(role, HttpStatus.OK);
+        }
+		return new ResponseEntity<>(new CustomErrorType("Role with name " + name 
+        + " not found"), HttpStatus.NOT_FOUND);	
     }
-
-    @DeleteMapping("/role/{name}")
-    public boolean deleteRole(@PathVariable String name){
-       roleService.delete(name);
-       return true;
-    }
-
-    @PutMapping("/role")
-    public Role updateRole(Role role){
-        return roleService.save(role);
-    }
-
+    
+    //Create a role
     @PostMapping("/role")
-    public Role createRole(Role role){
-        return roleService.save(role);
+	public ResponseEntity<?> createRole(@RequestBody Role role, UriComponentsBuilder ucBuilder) {
+        String name = role.getName();
+		if (roleService.isRoleExist(name)) {
+			return new ResponseEntity<>(new CustomErrorType("Unable to create. A Role with name " + 
+			name + " already exist."), HttpStatus.CONFLICT);
+		}
+		roleService.save(role);
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.setLocation(ucBuilder.path("/api/role/{name}").buildAndExpand(role.getName()).toUri());
+		return new ResponseEntity<String>(headers, HttpStatus.CREATED);
+	}
+
+    //Update a role
+    @PutMapping("/role/{name}")
+	public ResponseEntity<?> updateRole(@PathVariable("name") String name, @RequestBody Role role) {
+        if(roleService.isRoleExist(name)){
+            Role currentRole = roleService.get(name);
+            currentRole.setName(role.getName());
+            roleService.save(currentRole);
+
+            return new ResponseEntity<Role>(currentRole, HttpStatus.OK);    
+        }
+
+        return new ResponseEntity<>(new CustomErrorType("Unable to upate. Role with name " + name + " not found."),
+        HttpStatus.NOT_FOUND);
     }
+
+    //Delete a role
+    @DeleteMapping("/role/{name}")
+	public ResponseEntity<?> deleteRole(@PathVariable("name") String name) {	     
+        if(roleService.isRoleExist(name)){
+            roleService.delete(name);
+            return new ResponseEntity<Role>(HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(new CustomErrorType("Unable to delete. Role with name " + name + " not found."),
+					HttpStatus.NOT_FOUND);
+    }
+    
+    //Delete all roles
+    @DeleteMapping("/role")
+	public ResponseEntity<Role> deleteAllRoles() {
+		roleService.deleteAll();
+		return new ResponseEntity<Role>(HttpStatus.NO_CONTENT);
+	}
 }
